Use `u64` timestamps

+++ src/sampler.rs
@@ -1,6 +1,6 @@
 use std::sync::mpsc;
 
 pub trait Sampler {
-    fn sampled_channel(&self) -> &mpsc::Receiver<(f64, f64)>;
+    fn sampled_channel(&self) -> &mpsc::Receiver<(u64, f64)>;
     fn stop(self: Box<Self>);
 }
